fix:
(1) JoinBoard 와 Member 사이를 @OneToOne 에서 @ManyToOne으로 변경완료
=> 이유 : 처음 @OneToOne 으로 설정한 이유는 도배 방지를위해서였다.
그러나 soft delete 처리가 됐을때 @OneToOne 이면 새로운 글 작성이 불가능하게 돼버리기때문에 @ManyToOne 으로 수정하였다.

(2) 글에 Member가 (작성자가) 할당되도록 변경

(3) JoinBoard 작성, 수정, 삭제 할때 권한 로직 추가완료

(4) JoinBoard 테스트 코드 수정완료

+++ src/main/java/org/noteam/be/joinBoard/controller/JoinBoardController.java
@@ -1,21 +1,29 @@
 package org.noteam.be.joinBoard.controller;
 
 import lombok.RequiredArgsConstructor;
+import org.noteam.be.joinBoard.domain.JoinBoard;
 import org.noteam.be.joinBoard.dto.JoinBoardCardResponse;
 import org.noteam.be.joinBoard.dto.JoinBoardRegisterRequest;
 import org.noteam.be.joinBoard.dto.JoinBoardResponse;
 import org.noteam.be.joinBoard.dto.JoinBoardUpdateRequest;
 import org.noteam.be.joinBoard.service.JoinBoardService;
+import org.noteam.be.member.domain.Member;
+import org.noteam.be.member.domain.Role;
+import org.noteam.be.system.exception.ExceptionMessage;
+import org.noteam.be.system.exception.member.ExistingAuthenticationIsNull;
 import org.noteam.be.system.response.ResponseCode;
 import org.noteam.be.system.response.ResponseData;
+import org.noteam.be.system.util.SecurityUtil;
 import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
 
 @RestController
 @RequestMapping("/api/join-board")
 @RequiredArgsConstructor
 public class JoinBoardController {
+
 
     private final JoinBoardService joinBoardService;
 
@@ -23,10 +31,20 @@ public class JoinBoardController {
     @PostMapping
     public ResponseEntity<ResponseData<JoinBoardResponse>> createJoinBoard(@RequestBody JoinBoardRegisterRequest dto){
 
-        JoinBoardResponse res = joinBoardService.createJoinBoardByDto(dto);
+        Long currentMemberId = SecurityUtil.getCurrentMemberId();
+
+        if ( currentMemberId==null ) {
+            throw new ExistingAuthenticationIsNull(ExceptionMessage.MemberAuth.EXISTING_AUTHENTICATION_IS_NULL);
+        }
+
+        Member findMember = joinBoardService.findMemberById(currentMemberId);
+
+        JoinBoardResponse res = joinBoardService.createJoinBoardByDto(dto, findMember);
 
         return ResponseData.toResponseEntity(ResponseCode.POST_JOIN_BOARD_SUCCESS ,res);
     }
+
+
 
 
     @GetMapping
@@ -39,6 +57,8 @@ public class JoinBoardController {
     }
 
 
+
+
     @GetMapping("/{joinBoardId}")
     public ResponseEntity<ResponseData<JoinBoardResponse>> getJoinBoard(@PathVariable Long joinBoardId){
 
@@ -48,20 +68,57 @@ public class JoinBoardController {
     }
 
 
+
+
     @PutMapping("/{joinBoardId}")
     public ResponseEntity<ResponseData<JoinBoardResponse>> updateJoinBoard(@PathVariable Long joinBoardId, @RequestBody JoinBoardUpdateRequest dto) {
 
-        JoinBoardResponse res = joinBoardService.updateJoinBoardById(joinBoardId, dto);
+        Long currentMemberId = SecurityUtil.getCurrentMemberId();
+        if ( currentMemberId == null ) {
+            throw new ExistingAuthenticationIsNull(ExceptionMessage.MemberAuth.EXISTING_AUTHENTICATION_IS_NULL);
+        }
 
-        return ResponseData.toResponseEntity(ResponseCode.UPDATE_JOIN_BOARD_SUCCESS, res);
+        Member currentMember = joinBoardService.findMemberById(currentMemberId);
+
+        JoinBoard currentJoinBoard = joinBoardService.getJoinBoardEntityById(joinBoardId);
+
+
+        if ( currentJoinBoard.getMember().equals(currentMember) || currentMember.getRole().equals(Role.ADMIN) ){
+
+            JoinBoardResponse res = joinBoardService.updateJoinBoardById(joinBoardId, dto);
+
+            return ResponseData.toResponseEntity(ResponseCode.UPDATE_JOIN_BOARD_SUCCESS, res);
+
+        } else {
+            throw new ExistingAuthenticationIsNull(ExceptionMessage.MemberAuth.EXISTING_AUTHENTICATION_IS_NULL);
+        }
+
     }
+
 
     @DeleteMapping("/{joinBoardId}")
     public ResponseEntity<ResponseData> deleteJoinBoard(@PathVariable Long joinBoardId) {
 
-        joinBoardService.deleteJoinBoardById(joinBoardId);
+        Long currentMemberId = SecurityUtil.getCurrentMemberId();
+        if ( currentMemberId == null ) {
+            throw new ExistingAuthenticationIsNull(ExceptionMessage.MemberAuth.EXISTING_AUTHENTICATION_IS_NULL);
+        }
 
-        return ResponseData.toResponseEntity(ResponseCode.DELETE_JOIN_BOARD_SUCCESS);
+        Member currentMember = joinBoardService.findMemberById(currentMemberId);
+
+        JoinBoard currentJoinBoard = joinBoardService.getJoinBoardEntityById(joinBoardId);
+
+
+        if ( currentJoinBoard.getMember().equals(currentMember) || currentMember.getRole().equals(Role.ADMIN) ){
+
+            joinBoardService.deleteJoinBoardById(joinBoardId);
+
+            return ResponseData.toResponseEntity(ResponseCode.DELETE_JOIN_BOARD_SUCCESS);
+
+        } else {
+            throw new ExistingAuthenticationIsNull(ExceptionMessage.MemberAuth.EXISTING_AUTHENTICATION_IS_NULL);
+        }
+
     }
